requirements.txt
alive_progress==1.5.1
biopython==1.73
cobra==0.16.0
depinfo==1.4.0
future==0.16.0
matplotlib==3.2.1
memote==0.12.0
numpy==1.18.2
optlang==1.4.3
pandas==1.0.0
pip==20.0.2
PubChemPy==1.0.4
python==3.7.4
python-libsbml-experimental==5.18.0
ruamel.yaml==0.16.5
setuptools==45.2.0
six==1.14.0
swiglpk==4.65.0
wheel==0.34.2
